fix: split confirm-gate module under the loc cap and address review nits
Fixes the CI R8 (oversized-module) failure plus @claude review advisories.

- split agent/controller.rs (was 1676 LOC, over the 1400 R8 hard cap): moved
  the confirm-gate execution + validation (resolve_write, run_write_raced,
  validate_edited_args, is_routing_egress_key, clamp_json_strings,
  WriteResolution, CONFIRM_TIMEOUT/ARGS_DISPLAY_CAP) and their tests into
  agent/gate.rs. Behavior-preserving; controller.rs now 985, gate.rs 1080,
  both under the cap.
- panel: clear a stale confirm card when a non-confirm step resumes the run,
  which happens on the 300s server-side confirm timeout (deny-and-resume).
- validate_edited_args: reject routing/egress keys at ANY nesting depth
  (recursive scan), so a future object-typed gated-tool arg can't smuggle a
  nested provider/base_url/job_id.
- docs: correct the callId format to {step}-{idx}-{tool} in the three doc
  comments that still said {step}-{tool}.

cargo build + clippy (-D warnings) + fmt clean; typecheck 12/12; 3078 TS tests
pass. Rust tests run in CI.
